line2Lerp - reorg
