+++ line2Lerp.html
@@ -12,7 +12,7 @@
             background-color: #000;
             font-family: Arial, sans-serif;
         }
-        
+
         #info {
             position: absolute;
             top: 20px;
@@ -24,17 +24,17 @@
             max-width: 300px;
             z-index: 1000;
         }
-        
+
         #info h2 {
             margin-top: 0;
             color: #00ff88;
         }
-        
+
         #info p {
             margin: 10px 0;
             line-height: 1.4;
         }
-        
+
         .highlight {
             color: #00ff88;
             font-weight: bold;
@@ -48,7 +48,7 @@
         <p>The morph target is the same spline with <span class="highlight">sine displacement set to 0</span> (straight line).</p>
         <p>Use the slider to morph between the sine curve and straight line.</p>
     </div>
-    
-    <script type="module" src="./src/line2-lerp.js"></script>
+
+    <script type="module" src="src/line2Lerp.js"></script>
 </body>
 </html>
